Pull help rendering, the train guard and shared options out of the CLI
  cli/rendering.py  the rich-click styling, the two command panels, the column-alignment patch
  cli/guards.py     require_train — the refusal six commands make before doing any work
  cli/options.py    the four option declarations that are byte-identical across commands

`configure_help()` replaces the import-time side effects, and `cli/__init__.py` calls it through
`create_olinda_cli()`. It now also carries the idempotence guard the old code lacked: the alignment
patch wraps RichCommandPanel.get_table, so calling it twice wrapped the wrapper. Verified idempotent
across three calls.

options.py holds only `--task`, `--val-frac`, `--soft-smiles-column` and `--hard-smiles-column`, whose
declarations are identical everywhere they appear. The nine flags that differ stay with their commands:
`--model-dir` is declared by five commands with five different descriptions, and help text is the
command's interface, so those are five commands explaining what they will do with the directory rather
than duplication to collapse.

Help output is identical across all eleven screens.
